requirements.txt
azure-functions==1.21.0
requests==2.32.3
azure-search-documents==11.5.2
azure-core==1.32.0
azure-identity==1.19.0
azure-servicebus==7.12.3
azure-ai-projects==1.0.0b11
azure-ai-agents==1.0.0
